Use dummy keypair generation in test-utils
Moving the `derive_dummy_keypair` to test-utils.
