+++ air/tests/test_module/features/signatures.rs
@@ -14,8 +14,10 @@
  * limitations under the License.
  */
 
-use air_interpreter_signatures::{derive_dummy_keypair, SignatureTracker};
-use air_test_framework::{ephemeral::PeerId, AirScriptExecutor};
+use air_interpreter_signatures::SignatureTracker;
+use air_test_framework::ephemeral::PeerId;
+use air_test_framework::AirScriptExecutor;
+use air_test_utils::key_utils::derive_dummy_keypair;
 use air_test_utils::prelude::*;
 use air_test_utils::test_runner::TestRunParameters;
