+++ crates/air-lib/interpreter-signatures/Cargo.toml
@@ -17,9 +17,4 @@ fluence-keypair = "0.10.1"
 serde = { version = "1.0.159", features = ["derive"] }
 serde_json = "1.0.96"
 
-sha2 = "0.10.6"
 bs58 = "0.4.0"
-
-# TODO to be moved to test-utils:
-ed25519-dalek = "1.0.1"
-rand_chacha = "0.2.2"
