+++ crates/air-lib/test-utils/Cargo.toml
@@ -31,6 +31,10 @@ serde_json = "1.0.95"
 serde = { version = "1.0.159", features = ["derive"] }
 fluence-keypair = "0.10.1"
 
+ed25519-dalek = "1.0.1"
+rand_chacha = "0.2.2"
+sha2 = "0.10.6"
+
 [dev-dependencies]
 maplit = "1.0.2"
