+++ crates/air-lib/test-utils/src/lib.rs
@@ -28,6 +28,7 @@
 
 pub mod call_services;
 pub mod executed_state;
+pub mod key_utils;
 pub mod test_runner;
 
 #[cfg(feature = "test_with_native_code")]
